Switch size of byte content to quint32
UnknownContent::addAlternateDragObjects() tries to read the binary data
of "unknown content" notes. The problem here is that the size for the
data of each MIME format is read as quint64, whereas
NoteFactory::createNoteUnknown() saves it as quint32, hence not reading
the size correctly (and then crashing because the size is a very big
value).

Since the on-disk data is stored using quint32, switch the reading to
quint32 too.

+++ src/notecontent.cpp
@@ -2529,7 +2529,7 @@ void UnknownContent::addAlternateDragObjects(QMimeData *dragObject)
             }
         } while (!line.isEmpty() && !stream.atEnd());
         // Add the streams:
-        quint64 size; // TODO: It was quint32 in version 0.5.0 !
+        quint32 size;
         QByteArray *array;
         for (int i = 0; i < mimes.count(); ++i) {
             // Get the size:
